pyproject.toml
[project]
name = "qa-guru-hw"
version = "0.1.0"
description = ""
authors = [
    {name = "Anton Yaroslavtsev",email = "a.yaroslavtsev@rt-solar.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pytest (==8.4.1)",
    "python-dotenv (==1.1.1)",
    "allure-pytest (==2.15.0)",
    "selene (==2.0.0rc9)",
    "requests (==2.32.4)",
    "pydantic (==2.12.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pylint (>=3.3.9,<4.0.0)"
]
